requirements.txt
django
requests
python-dotenv
psycopg2-binary
djangorestframework
djangorestframework_simplejwt